Dealt with float comparison problem in data_utils.normalization
When image pixel values were normalized to -1 - 1 using
data_utils.normalize and then inverse normalized to 0 - 1 with
data_utils.inverse_normalize, it was possible to end up with out
out of range floats (e.g. 1.000001), which would cause matplotlib
(which expects floats between 0 and 1) to throw an error.
Added a check for this in normalize and inverse_normalize.